opt-in for the package cleanup and build because EnvDTE operations in powershell do not block

+++ src/CreateReleasePackage/tools/visualstudio.psm1
@@ -22,16 +22,30 @@ function New-Release {
     $outputDir =  $activeConfiguration.Properties.Item("OutputPath").Value
 
     $buildDir = Join-Path $projectDir $outputDir
-    if (Test-Path $buildDir) {
-        Write-Message "Clearing existing nupkg files from folder $outputDir"
-        Remove-Item "$buildDir\*.nupkg"
-    } else {
-        Write-Message "Build output folder $buildDir does not exist, skipping"
+
+    # because the EnvDTE build operations doesn't block on Win7
+    # we're not going to clean up packages here because
+    # we expect you to be a responsible adult
+    # with your artifacts
+
+    if ($psversiontable.psversion.major -gt 2) {
+
+        if (Test-Path $buildDir) {
+            Write-Message "Clearing existing nupkg files from folder $outputDir"
+            Remove-Item "$buildDir\*.nupkg"
+        } else {
+            Write-Message "Build output folder $buildDir does not exist, skipping"
+        }
+
+        Write-Message "Building project $ProjectName"
+
+        $dte.Solution.SolutionBuild.Clean($true)
+
+        $dte.Solution.SolutionBuild.BuildProject( `
+            $activeConfiguration.ConfigurationName, `
+            $project.FullName, `
+            $true)
     }
-
-    Write-Message "Building project $ProjectName"
-
-    $dte.Solution.SolutionBuild.BuildProject($activeConfiguration.ConfigurationName, $project.FullName, $true)
 
     Write-Message "Publishing release for project $ProjectName"
